Fix StepThreeFormMintedCapped test
