+++ src/components/StepThree/index.js
@@ -7,11 +7,10 @@ import { Loader } from '../Common/Loader'
 import { CHAINS, NAVIGATION_STEPS } from '../../utils/constants'
 import { StepNavigation } from '../Common/StepNavigation'
 import { checkWeb3, getNetWorkNameById, getNetworkVersion } from '../../utils/blockchainHelpers'
-import { getStep3Component } from './utils'
+import { getStep3Component, tierDurationUpdater } from './utils'
 import { inject, observer } from 'mobx-react'
 import { noGasPriceAvailable, warningOnMainnetAlert } from '../../utils/alerts'
 import { navigateTo } from '../../utils/utils'
-import createDecorator from 'final-form-calculate'
 
 const logger = logdown('TW:StepThree')
 const { CROWDSALE_SETUP } = NAVIGATION_STEPS
@@ -140,20 +139,7 @@ export class StepThree extends Component {
       })
   }
 
-  calculator = createDecorator({
-    field: /.+\.endTime/,
-    updates: (value, name) => {
-      const nextTierIndex = +name.match(/(\d+)/)[1] + 1
-      const { tierStore } = this.props
-      const newValue = {}
-
-      if (tierStore.tiers[nextTierIndex]) {
-        newValue[`tiers[${nextTierIndex}].startTime`] = value
-      }
-
-      return newValue
-    }
-  })
+  calculator = tierDurationUpdater(this.props.tierStore.tiers)
 
   render() {
     if (this.state.initialTiers.length === 0) {
