+++ src/components/StepThree/utils.js
@@ -2,6 +2,7 @@ import moment from 'moment'
 import { StepThreeFormMintedCapped } from './StepThreeFormMintedCapped'
 import { StepThreeFormDutchAuction } from './StepThreeFormDutchAuction'
 import { CROWDSALE_STRATEGIES } from '../../utils/constants'
+import createDecorator from 'final-form-calculate'
 
 export function defaultCompanyStartDate() {
   return moment()
@@ -25,4 +26,20 @@ export const getStep3Component = strategy => {
     default:
       return StepThreeFormMintedCapped
   }
+}
+
+export const tierDurationUpdater = tiers => {
+  return createDecorator({
+    field: /.+\.endTime/,
+    updates: (value, name) => {
+      const nextTierIndex = +name.match(/(\d+)/)[1] + 1
+      const newValue = {}
+
+      if (tiers[nextTierIndex]) {
+        newValue[`tiers[${nextTierIndex}].startTime`] = value
+      }
+
+      return newValue
+    }
+  })
 }
